Fix example tag name in README
Tag names are case-insensitive in Docker, and Docker 1.10 introduced
validation logic that causes `docker build` to fail if the specified tag
contains uppercase characters, as stated in https://github.com/docker/distribution/issues/1433.

It appears that Docker container names are case-sensitive, but the
default container names are all-lowercase with underscores, and I think
it looks better to use the same naming convention for tag names and
container names in the `docker run` example command.

I chose hyphens over underscores because our tags in this repo use
hyphens (e.g. 0.0.1-alpha-onbuild), but underscores would be fine too.